+++ age/src/lib.rs
@@ -149,8 +149,8 @@ mod primitives;
 mod protocol;
 mod util;
 
-pub use error::{DecryptError, EncryptError};
-pub use identity::{IdentityFile, IdentityFileEntry};
+pub use error::{DecryptError, EncryptError, IdentityFileConvertError};
+pub use identity::IdentityFile;
 pub use primitives::stream;
 pub use protocol::{Decryptor, Encryptor};
 
@@ -278,6 +278,9 @@ pub trait Callbacks: Clone + Send + Sync + 'static {
     ///
     /// This can be used to prompt the user to take some physical action, such as
     /// inserting a hardware key.
+    ///
+    /// No guarantee is provided that the user sees this message (for example, if there is
+    /// no UI for displaying messages).
     fn display_message(&self, message: &str);
 
     /// Requests that the user provides confirmation for some action.
@@ -300,10 +303,43 @@ pub trait Callbacks: Clone + Send + Sync + 'static {
     /// Requests non-private input from the user.
     ///
     /// To request private inputs, use [`Callbacks::request_passphrase`].
+    ///
+    /// Returns:
+    /// - `Some(input)` with the user-provided input.
+    /// - `None` if no input could be requested from the user (for example, if there is no
+    ///   UI for displaying messages or typing inputs).
     fn request_public_string(&self, description: &str) -> Option<String>;
 
     /// Requests a passphrase to decrypt a key.
+    ///
+    /// Returns:
+    /// - `Some(passphrase)` with the user-provided passphrase.
+    /// - `None` if no passphrase could be requested from the user (for example, if there
+    ///   is no UI for displaying messages or typing inputs).
     fn request_passphrase(&self, description: &str) -> Option<SecretString>;
+}
+
+/// An implementation of [`Callbacks`] that does not allow callbacks.
+///
+/// No user interaction will occur; [`Recipient`] or [`Identity`] implementations will
+/// receive `None` from the callbacks that return responses, and will act accordingly.
+#[derive(Clone, Copy, Debug)]
+pub struct NoCallbacks;
+
+impl Callbacks for NoCallbacks {
+    fn display_message(&self, _: &str) {}
+
+    fn confirm(&self, _: &str, _: &str, _: Option<&str>) -> Option<bool> {
+        None
+    }
+
+    fn request_public_string(&self, _: &str) -> Option<String> {
+        None
+    }
+
+    fn request_passphrase(&self, _: &str) -> Option<SecretString> {
+        None
+    }
 }
 
 /// Helper for fuzzing the Header parser and serializer.